// backend/src/models/condicaoPagto.ts
export interface CondicaoPagto {
  id: number;
  desfin: string;
  przmed?: number;
  qtdpar?: string;
  parc01?: string; parc02?: string; parc03?: string; parc04?: string;
  parc05?: string; parc06?: string; parc07?: string; parc08?: string;
  parc09?: string; parc10?: string; parc11?: string; parc12?: string;
  parc13?: string; parc14?: string; parc15?: string;
  salpar?: string;
  avista?: string;
  entrad?: string;
  invert?: string;
  safrax?: string;
  descon?: string;
  entradaSafra?: number;
  descontoEntradaSafra?: number;
  descontoSaldoSafra?: number;
  saldoSafra?: number;
}